+++ backend/config/routes/api_admin.php
@@ -203,6 +203,26 @@ return static function (RouteCollectorProxy $group) {
                     );
 
                     $group->get(
+                        '/rsas',
+                        Controller\Api\Admin\Rsas\GetAction::class
+                    )->setName('api:admin:rsas');
+
+                    $group->post(
+                        '/rsas',
+                        Controller\Api\Admin\Rsas\PostAction::class
+                    );
+
+                    $group->post(
+                        '/rsas/license',
+                        Controller\Api\Admin\Rsas\PostLicenseAction::class
+                    )->setName('api:admin:rsas-license');
+
+                    $group->delete(
+                        '/rsas/license',
+                        Controller\Api\Admin\Rsas\DeleteLicenseAction::class
+                    );
+
+                    $group->get(
                         '/shoutcast',
                         Controller\Api\Admin\Shoutcast\GetAction::class
                     )->setName('api:admin:shoutcast');