+++ backend/src/Radio/Enums/FrontendAdapters.php
@@ -6,12 +6,14 @@ namespace App\Radio\Enums;
 
 use App\Radio\Frontend\AbstractFrontend;
 use App\Radio\Frontend\Icecast;
+use App\Radio\Frontend\Rsas;
 use App\Radio\Frontend\Shoutcast;
 
 enum FrontendAdapters: string implements AdapterTypeInterface
 {
     case Icecast = 'icecast';
     case Shoutcast = 'shoutcast2';
+    case Rsas = 'rsas';
     case Remote = 'remote';
 
     public function getValue(): string
@@ -24,6 +26,7 @@ enum FrontendAdapters: string implements AdapterTypeInterface
         return match ($this) {
             self::Icecast => 'Icecast 2.4',
             self::Shoutcast => 'Shoutcast DNAS 2',
+            self::Rsas => 'Rocket Streaming Audio Server (RSAS)',
             self::Remote => 'Remote',
         };
     }
@@ -36,6 +39,7 @@ enum FrontendAdapters: string implements AdapterTypeInterface
         return match ($this) {
             self::Icecast => Icecast::class,
             self::Shoutcast => Shoutcast::class,
+            self::Rsas => Rsas::class,
             default => null
         };
     }
@@ -48,14 +52,15 @@ enum FrontendAdapters: string implements AdapterTypeInterface
     public function supportsMounts(): bool
     {
         return match ($this) {
-            self::Shoutcast, self::Icecast => true,
+            self::Shoutcast, self::Icecast, self::Rsas => true,
             default => false
         };
     }
 
     public function supportsReload(): bool
     {
-        return self::Icecast === $this;
+        return self::Icecast === $this
+            || self::Rsas === $this;
     }
 
     public static function default(): self