+++ frontend/components/Admin/Stations/stationFormProps.ts
@@ -2,6 +2,10 @@ export default {
     // Profile
     timezones: Object,
     // Frontend
+    isRsasInstalled: {
+        type: Boolean,
+        default: false
+    },
     isShoutcastInstalled: {
         type: Boolean,
         default: false