+++ frontend/components/Admin/routes.ts
@@ -98,6 +98,11 @@ export default function useAdminRoutes(): RouteRecordRaw[] {
             name: 'admin:install_shoutcast:index'
         },
         {
+            path: '/install_rsas',
+            component: () => import('~/components/Admin/Rsas.vue'),
+            name: 'admin:install_rsas:index'
+        },
+        {
             path: '/install_stereo_tool',
             component: () => import('~/components/Admin/StereoTool.vue'),
             name: 'admin:stereo_tool:index'